# full_openapi.yaml
...
openapi: 3.1.0
info:
  title: IMT Second-Hand Marketplace — Tier A (Minimum)
  version: 1.0.0
  description: |
    Teacher-provided API for automated testing. All strings and test
    descriptions are in English. Students MUST implement this spec as-is.
    Any deviation may cause tests to fail.
servers:
  - url: http://127.0.0.1:5050
    description: Local development server
...
components:
  securitySchemes:
    testUserHeader:
      type: apiKey
      in: header
      name: X-User-Email
      description: |
        For automated tests only. If an endpoint requires authentication,
        the client MUST send this header with a valid user email. To act
        as admin, use the exact value `admin@imt.test`.
  parameters:
    CategoryId:
      name: category_id
      in: path
      required: true
      description: Numeric identifier of a category (>= 1).
      schema: { type: integer, minimum: 1 }
    AddressId:
      name: address_id
      in: path
      required: true
      description: Numeric identifier of a user-owned address (>= 1).
      schema: { type: integer, minimum: 1 }
...
  responses:
    UnauthorizedError:
      description: |
        Missing or invalid authentication for an endpoint that requires it.
        The server MUST NOT expose whether a given email exists.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          examples:
            missingHeader:
              summary: Missing X-User-Email header
              value: { error: "Unauthorized: missing X-User-Email" }
    ForbiddenError:
      description: Authenticated but not permitted to perform the operation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          examples:
            notOwner:
              value: { error: "Forbidden: not owner" }
            notAdmin:
              value: { error: "Forbidden: admin only" }
    NotFoundError:
      description: The requested resource does not exist or is not visible.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          examples:
            generic:
              value: { error: "Not found" }
    ConflictError:
      description: Request conflicts with the current state (e.g., double action, constraints).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          examples:
            alreadyTerminal:
              value: { error: "Conflict: purchase is already terminal" }
            categoryHasChildren:
              value: { error: "Conflict: category has children or listings" }
...
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message in English.
    Email:
      type: string
      format: email
      description: A syntactically valid email address (RFC 5322 local@domain).
      examples: [ "user@example.com" ]
    MoneyCents:
      type: integer
      minimum: 0
      description: |
        Currency amount in **euro cents**. Always non-negative in responses.
        Negative values may appear only as transaction `amount_cents` for
        debits (e.g., purchases).
...
    User:
      type: object
      description: Minimal user representation.
      required: [email, credits_cents]
      properties:
        email: { $ref: '#/components/schemas/Email' }
        credits_cents:
          $ref: '#/components/schemas/MoneyCents'
      examples:
        - email: "alice@example.com"
          credits_cents: 3500
    Address:
      type: object
      description: |
        Delivery address owned by a specific user. Address IDs are unique
        per server (not per user).
      required: [id, line1, city, postal_code]
      properties:
        id:
          type: integer
          minimum: 1
          description: Server-assigned unique numeric identifier.
        line1: { type: string, minLength: 1, maxLength: 200 }
        line2: { type: [string, "null"], maxLength: 200 }
        city:  { type: string, minLength: 1, maxLength: 100 }
        postal_code:
          type: string
          description: French postal code (no strict format enforced here).
          minLength: 1
          maxLength: 20
      examples:
        - id: 42
          line1: "10 Rue de la Paix"
          line2: null
          city: "Nantes"
          postal_code: "44000"
    Category:
      type: object
      description: |
        Category node in an infinitely nested tree. Use `parent_id = null`
        for root categories.
      required: [id, name]
      properties:
        id: { type: integer, minimum: 1 }
        name: { type: string, minLength: 1, maxLength: 120 }
        parent_id:
          type: [integer, "null"]
          minimum: 1
          description: Parent category id, or null for root.
      examples:
        - id: 7
          name: "Sneakers"
          parent_id: 3
...
security:
  - testUserHeader: []
paths:
...
  /api/categories:
    get:
      tags: [categories]
      operationId: listCategories
      summary: List all categories as a flat collection
      description: |
        Returns **all** categories as a flat list. Each category includes
        its `parent_id` (or `null` for roots). Clients can rebuild the tree
        by following `parent_id` links.

        Notes:
        - IDs are stable and unique.
        - There is no pagination (category counts are expected to be small).
      responses:
        '200':
          description: Array of categories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Category' }
              examples:
                sample:
                  value:
                    - { id: 1, name: "Clothing", parent_id: null }
                    - { id: 2, name: "Shoes", parent_id: 1 }
                    - { id: 7, name: "Sneakers", parent_id: 2 }

    post:
      tags: [categories]
      operationId: createCategory
      summary: Create a category (admin only)
      description: |
        Creates a new category. Caller MUST be admin. `parent_id` may be
        omitted or null for a root node. Category names need not be unique
        globally (but keeping them distinct is recommended).
      security: [ { testUserHeader: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 120 }
                parent_id:
                  type: [integer, "null"]
                  minimum: 1
            examples:
              root:
                value: { name: "Accessories", parent_id: null }
              child:
                value: { name: "Sneakers", parent_id: 2 }
      responses:
        '201':
          description: Created category
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
  /api/categories/{category_id}:
    delete:
      tags: [categories]
      operationId: deleteCategory
      summary: Delete a category (admin only)
      description: |
        Deletes the category if and only if it has **no child categories**
        and **no listings** assigned (directly or through descendants).
        Otherwise, the server MUST return 409 Conflict.
      security: [ { testUserHeader: [] } ]
      parameters: [ { $ref: '#/components/parameters/CategoryId' } ]
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { $ref: '#/components/responses/ConflictError' }

  # --------------------------------------------------------------------
  # BUYER-PROTECTION CONFIG (admin read/write, public read)
  # --------------------------------------------------------------------
...
  /api/users:
    post:
      tags: [users]
      operationId: registerUser
      summary: Register a user account (email + password)
      description: |
        Creates a new user. The email is the unique account identifier.
        No email verification is performed. Password change endpoints are
        out of scope for Tier A.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { $ref: '#/components/schemas/Email' }
                password:
                  type: string
                  minLength: 4
                  maxLength: 200
                  description: Stored server-side as a salted hash.
            examples:
              ok:
                value: { email: "alice@example.com", password: "secret" }
      responses:
        '201':
          description: Created user (initial credits are zero)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
              examples:
                ok:
                  value: { email: "alice@example.com", credits_cents: 0 }
        '409':
          description: Email already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
              examples:
                dup:
                  value: { error: "Conflict: email already exists" }
...
  /api/addresses:
    get:
      tags: [addresses]
      operationId: listMyAddresses
      summary: List the authenticated user’s addresses
      security: [ { testUserHeader: [] } ]
      responses:
        '200':
          description: Array of addresses owned by the caller
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Address' }
              examples:
                sample:
                  value:
                    - { id: 77, line1: "5 Avenue Anatole", line2: null, city: "Paris", postal_code: "75007" }

    post:
      tags: [addresses]
      operationId: addAddress
      summary: Add a new address to the authenticated user
      security: [ { testUserHeader: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              # Note: id is server-assigned; client must NOT send it.
              type: object
              required: [line1, city, postal_code]
              properties:
                line1: { type: string, minLength: 1, maxLength: 200 }
                line2: { type: [string, "null"], maxLength: 200 }
                city:  { type: string, minLength: 1, maxLength: 100 }
                postal_code: { type: string, minLength: 1, maxLength: 20 }
            examples:
              ok:
                value: { line1: "10 Rue de la Paix", line2: null, city: "Nantes", postal_code: "44000" }
      responses:
        '201':
          description: Created address
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Address' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /api/addresses/{address_id}:
    put:
      tags: [addresses]
      operationId: replaceAddress
      summary: Replace (update) an address owned by the authenticated user
      description: |
        Full update. The request body MUST include all mutable fields
        (line1, line2, city, postal_code). The `id` in the body, if present,
        MUST be ignored by the server.
      security: [ { testUserHeader: [] } ]
      parameters: [ { $ref: '#/components/parameters/AddressId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [line1, city, postal_code]
              properties:
                line1: { type: string, minLength: 1, maxLength: 200 }
                line2: { type: [string, "null"], maxLength: 200 }
                city:  { type: string, minLength: 1, maxLength: 100 }
                postal_code: { type: string, minLength: 1, maxLength: 20 }
            examples:
              replaceAll:
                value: { line1: "99 Quai de Jemmapes", line2: "Bât A", city: "Paris", postal_code: "75010" }
      responses:
        '200':
          description: Updated address
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Address' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }

    delete:
      tags: [addresses]
      operationId: deleteAddress
      summary: Delete an address owned by the authenticated user
      security: [ { testUserHeader: [] } ]
      parameters: [ { $ref: '#/components/parameters/AddressId' } ]
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }

  # --------------------------------------------------------------------
  # LISTINGS (seller CRUD; public GET)
  # --------------------------------------------------------------------
...